requirements.txt
Django==1.10.4
Pillow==3.4.2
argparse==1.2.1
cloudinary==1.5.0
django-allauth==0.29.0
django-bootstrap-form==3.2.1
django-crispy-forms==1.6.1
django-flat-responsive==1.1.0
django-formtools==1.0
django-localflavor==1.3
django-simple-captcha==0.5.3
django-widget-tweaks==1.4.1
gunicorn==19.6.0
httplib2==0.9.2
oauthlib==2.0.1
python-openid==2.2.5
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
urllib3==1.19.1
whitenoise==3.2.2
distribute==0.6.27
dj-database-url==0.4.1
poster==0.8.1
psycopg2==2.4.5
wsgiref==0.1.2
